FIX fish position data type

+++ Fish.cpp
@@ -8,23 +8,23 @@ using namespace std;
 
 Fish::Fish()
 {
-    this->xPos = rand() % SCREEN_WIDTH + 1;
-    this->yPos = rand() % SCREEN_HEIGHT + 1;
+    this->xPos = (double)(rand() % SCREEN_WIDTH + 1);
+    this->yPos = (double)(rand() % SCREEN_HEIGHT + 1);
     this->foodEaten = 0;
     this->direction = rand()%2 == 0? true:false;
     this->hungerTime = 100;
 }
-
-void Fish::move(int dx, int dy){
+)
+void Fish::move(double dx, double dy){
     this->xPos += dx;
     this->yPos += dy;
 }
 
-int Fish::getXPos() const{
+double Fish::getXPos() const{
     return this->xPos;
 }
 
-int Fish::getYPos() const{
+double Fish::getYPos() const{
     return this->yPos;
 }
 bool Fish::getDirection() const{
